requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.13
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
botocore==1.20.43
boto3==1.17.43
django-storages==1.12.1